+++ docs/CHANGELOG.md
@@ -25,6 +25,7 @@ The entire project (i.e., the GraphQL API, REST API, and server) uses the same v
 - Return a `Boolean!` instead of a `Placeholder!` from `Mutation.unblockUser` to indicate whether the user got unblocked.
 - Use GitHub Flavored Markdown instead of CommonMark (update `scalar Bio`, `scalar GroupChatDescription`, and `scalar MessageText` accordingly).
 - Paginate `Query.readStars`.
+- Paginate `Query.searchMessages`.
 
 ### Removed
