+++ src/main/kotlin/graphql/routing/Models.kt
@@ -1158,7 +1158,9 @@ object CreatedSubscription :
     val placeholder = Placeholder
 }
 
-data class ChatMessages(val chat: Chat, val messages: List<MessageEdge>)
+data class ChatMessagesConnection(val edges: List<ChatMessagesEdge>, val pageInfo: PageInfo)
+
+data class ChatMessagesEdge(val chat: Chat, val messages: List<MessageEdge>)
 
 data class AccountsConnection(val edges: List<AccountEdge>, val pageInfo: PageInfo) {
     companion object {
